restructured views for homepage
