+++ views/partials/allBooks.handlebars
@@ -1,0 +1,8 @@
+<div>
+  <div>
+  <a href="/books/{{id}}"><h2>{{title}}</h2><a/>
+  <h3>Author: {{author}}</h3>
+  <h3>Genre: {{genre}}</h3>
+  <h5>Posted by {{user.name}}</h5>
+  </div>
+</div>
